pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "bm2-evaluator"
version = "0.1.0"
description = "Standalone evaluator for protein binder designs"
requires-python = ">=3.10"
dependencies = [
    "numpy>=1.24",
    "biopython>=1.82",
    "gemmi>=0.6",
]

[project.optional-dependencies]
sasa = ["freesasa>=2.2"]
dev = ["pytest>=7.4", "pytest-cov", "pyyaml>=6.0", "click>=8.0"]

[project.scripts]
bm2-eval = "bm2_evaluator.cli:main"
bm2 = "bm2.cli.main:main"

[tool.setuptools.packages.find]
include = ["bm2_evaluator*", "bm2*"]

[tool.pytest.ini_options]
testpaths = ["tests", "bm2/tests"]
